skills/requirements.txt
RPi.GPIO
smbus2
hermes-python>=0.1